+++ e2e/__tests__/toMatchSnapshotWithStringSerializer.test.ts
@@ -5,6 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import * as fs from 'fs';
 import * as path from 'path';
 import {cleanup, makeTemplate, writeFiles} from '../Utils';
 import runJest from '../runJest';
@@ -15,8 +16,17 @@ const DIR = path.resolve(
 );
 const TESTS_DIR = path.resolve(DIR, '__tests__');
 
+const readFile = filename =>
+  fs.readFileSync(path.join(TESTS_DIR, filename), 'utf8');
+
 beforeEach(() => cleanup(TESTS_DIR));
 afterAll(() => cleanup(TESTS_DIR));
+
+// Because the not written error might include Received,
+// match Snapshot as either diff annotation or concise label.
+const ORDINARY_FAILURE = /- Snapshot|Snapshot:/;
+
+const NOT_WRITTEN = 'not written'; // new snapshot with --ci option
 
 test('empty external', () => {
   // Make sure empty string as expected value of external snapshot
@@ -28,28 +38,80 @@ test('empty external', () => {
 
   {
     writeFiles(TESTS_DIR, {
-      [filename]: template(['""']), // empty string
+      [filename]: template([`''`]),
     });
-    const {stderr, status} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
     expect(stderr).toMatch('1 snapshot written from 1 test suite.');
-    expect(status).toBe(0);
+    expect(exitCode).toBe(0);
   }
 
   {
-    const {stderr, status} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
     expect(stderr).toMatch('Snapshots:   1 passed, 1 total');
     expect(stderr).not.toMatch('1 snapshot written from 1 test suite.');
-    expect(status).toBe(0);
+    expect(exitCode).toBe(0);
   }
 
   {
     writeFiles(TESTS_DIR, {
-      [filename]: template(['"non-empty"']),
+      [filename]: template([`'non-empty'`]),
     });
-    const {stderr, status} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
     expect(stderr).toMatch('Snapshots:   1 failed, 1 total');
     expect(stderr).not.toMatch('not written'); // not confused with --ci option
-    expect(stderr).toMatch(/- Snapshot|Snapshot:/); // ordinary report
-    expect(status).toBe(1);
+    expect(stderr).toMatch(ORDINARY_FAILURE);
+    expect(exitCode).toBe(1);
+  }
+});
+
+test('empty internal ci false', () => {
+  // Make sure empty string as expected value of internal snapshot
+  // is not confused with absence of snapshot.
+  const filename = 'empty-internal-ci-false.test.js';
+  const template = makeTemplate(
+    `test('string serializer', () => { expect($1).toMatchInlineSnapshot(); })`,
+  );
+
+  const received1 = `''`;
+  const received2 = `'non-empty'`;
+
+  {
+    writeFiles(TESTS_DIR, {
+      [filename]: template([received1]),
+    });
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    expect(stderr).toMatch('1 snapshot written from 1 test suite.');
+    expect(exitCode).toBe(0);
+  }
+
+  {
+    writeFiles(TESTS_DIR, {
+      [filename]: readFile(filename).replace(received1, received2),
+    });
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=false', filename]);
+    expect(stderr).toMatch('Snapshots:   1 failed, 1 total');
+    expect(stderr).not.toMatch('1 snapshot written from 1 test suite.');
+    expect(stderr).toMatch(ORDINARY_FAILURE);
+    expect(exitCode).toBe(1);
+  }
+});
+
+test('undefined internal ci true', () => {
+  // Make sure absence of internal snapshot
+  // is not confused with ordinary failure for empty string as expected value.
+  const filename = 'undefined-internal-ci-true.test.js';
+  const template = makeTemplate(
+    `test('explicit update', () => { expect($1).toMatchInlineSnapshot(); })`,
+  );
+
+  {
+    writeFiles(TESTS_DIR, {
+      [filename]: template([`'non-empty'`]),
+    });
+    const {stderr, exitCode} = runJest(DIR, ['-w=1', '--ci=true', filename]);
+    expect(stderr).toMatch('Snapshots:   1 failed, 1 total');
+    expect(stderr).not.toMatch(ORDINARY_FAILURE);
+    expect(stderr).toMatch(NOT_WRITTEN);
+    expect(exitCode).toBe(1);
   }
 });