+++ e2e/__tests__/watchModePatterns.test.ts
@@ -45,7 +45,7 @@ test('can press "p" to filter by file name', () => {
   const input = [{keys: ['p', 'b', 'a', 'r', '\r']}, {keys: ['q']}];
   setupFiles(input);
 
-  const {status, stdout, stderr} = runJest(DIR, [
+  const {exitCode, stdout, stderr} = runJest(DIR, [
     '--no-watchman',
     '--watchAll',
   ]);
@@ -58,14 +58,14 @@ test('can press "p" to filter by file name', () => {
     expect(wrap(rest)).toMatchSnapshot('test results');
     expect(wrap(summary)).toMatchSnapshot('test summary');
   });
-  expect(status).toBe(0);
+  expect(exitCode).toBe(0);
 });
 
 test('can press "t" to filter by test name', () => {
   const input = [{keys: ['t', '2', '\r']}, {keys: ['q']}];
   setupFiles(input);
 
-  const {status, stdout, stderr} = runJest(DIR, [
+  const {exitCode, stdout, stderr} = runJest(DIR, [
     '--no-watchman',
     '--watchAll',
   ]);
@@ -78,5 +78,5 @@ test('can press "t" to filter by test name', () => {
     expect(wrap(rest)).toMatchSnapshot('test results');
     expect(wrap(summary)).toMatchSnapshot('test summary');
   });
-  expect(status).toBe(0);
+  expect(exitCode).toBe(0);
 });